+++ inst/shiny/1/server.R
@@ -2,7 +2,7 @@
 # server.R
 
 shinyServer(function(input, output, session) {
-  # plot
+  ### plot ###
   
   plot_data <- reactive({
     # create a reactive plot_data object
@@ -12,7 +12,7 @@ shinyServer(function(input, output, session) {
     
     .color <- input$plot_color
     
-    plot_data <- data %>%
+    plot_data <- data1 %>%
       filter(color == .color) %>%
       mutate(cut = stringr::str_to_sentence(cut)) %>%
       group_by(cut, clarity, .drop = FALSE) %>%
@@ -28,15 +28,15 @@ shinyServer(function(input, output, session) {
   #   rownames = FALSE,
   #   options = list(pageLength = 5, scrollX = TRUE, lengthChange = FALSE)
   # )
-
+  
   plot_theme <- reactive({
     gg_theme(
+      gridlines = "vertical", 
       family  = "helvetica", 
-      size_title = ifelse(input$isMobile == FALSE, 11, 16), 
-      size_body = ifelse(input$isMobile == FALSE, 10, 15),
-      gridlines = "vertical" 
-      )
-  })
+      size_title = 11, 
+      size_body = 10 
+    )
+  }) 
   
   plot <- reactive({
     # create a reactive ggplot object
@@ -66,26 +66,88 @@ shinyServer(function(input, output, session) {
       x_labels = scales::comma_format(),
       col_labels = ggplot2::waiver(),
       title_wrap = title_wrap,
-      theme = plot_theme(), 
-      mobile = input$isMobile
+      theme = plot_theme()
     )
     
     return(plot)
-  }) %>%
-    bindCache(input$plot_color, input$isMobile)
+  })
   
   output$plot_desktop <- plotly::renderPlotly({
     plotly::ggplotly(plot(), tooltip = "text") %>%
       plotly_camera()
   }) 
   
-  output$plot_mobile <- renderPlot({
-    plot()
+  ### map ###
+  output$map <- leaflet::renderLeaflet({
+    
+    basemap %>%
+      leaflet::addMiniMap(
+        tiles = leaflet::providers$CartoDB.PositronNoLabels,
+        position = "bottomleft",
+        width = 100,
+        height = 100,
+        autoToggleDisplay = TRUE,
+        toggleDisplay = TRUE,
+        minimized = TRUE
+      )
+  })
+  
+  map_data <- reactive({
+    # create a reactive map_data object
+    
+    map_filter <- input$map_filter
+    
+    if (map_filter == "None") {
+      map_data <- data2
+    } else if (map_filter != "None") {
+      map_data <- data2 %>%
+        filter(trend_category == map_filter)
+    }
+    
+    return(map_data)
   }) 
+  
+  # output$map_data <- DT::renderDT(
+  #   map_data(), 
+  #   filter = "top",
+  #   rownames = FALSE,
+  #   options = list(pageLength = 5, scrollX = TRUE, lengthChange = FALSE)
+  # )
+  
+  draw_map <- function() {
+    # add leaflet code from make_data_vis.R
+    # change any placeholder character values to input widgets
+    # refer to a reactive map_data object as map_data()
+    # use reactive radius for points that get bigger as the user zooms in, if necessary
+    
+    size_reactive <- ifelse(input$map_zoom < 6, 1.5,
+             ifelse(input$map_zoom < 7, 2, 
+                    ifelse(input$map_zoom < 8, 3, 4)))
+    
+    title <- paste0("Monitored trends, 2008\u201317")
+    
+    leaf_sf_col(
+      map_data(),
+      col_var = trend_category,
+      size_point = size_reactive,
+      col_title = title
+    )
+  }
+  
+  observe({
+    req(map_data())
+    req(input$map_zoom) # wait for basemap before plotting.
+    
+    withProgress(message = "Loading", {
+      draw_map()
+    })
+  })
   
   ### table ###
   
   table_data <- reactive({
+    # create a reactive table_data object
+    
     ggplot2::diamonds %>%
       select(carat:price) %>%
       rename_with(snakecase::to_sentence_case)
@@ -95,9 +157,9 @@ shinyServer(function(input, output, session) {
     table_data(),
     filter = "top",
     rownames = FALSE,
-      options = list(pageLength = 10, scrollX = TRUE, lengthChange = FALSE)
-    )
-
+    options = list(pageLength = 10, scrollX = TRUE, lengthChange = FALSE)
+  )
+  
   ### download ###
   
   output$download <- downloadHandler(filename <- function() {
@@ -111,10 +173,10 @@ shinyServer(function(input, output, session) {
   ### download code ###
   
   output$download_code <- downloadHandler(filename <- function() {
-    "template1.zip"
+    "demo1.zip"
   },
   content <- function(file) {
-    file.copy("data/template1.zip", file)
+    file.copy("data/demo1.zip", file)
   },
   contentType = "application/zip")