+++ inst/shiny/2/global.R
@@ -6,10 +6,12 @@ library(dplyr)
 library(simplevis)
 
 # load data
+set.seed(123456789)
+
 data1 <- ggplot2::diamonds %>%
   slice_sample(prop = 0.1)
 
-data2 <- simplevis::example_sf_point %>%
+data2 <- simplevis::example_point %>%
   mutate(trend_category = factor(
     trend_category,
     levels = c("Improving", "Indeterminate", "Worsening")