+++ inst/shiny/zip_code.R
@@ -1,0 +1,6 @@
+zip::zipr(zipfile = "inst/shiny/1/data/demo1.zip", 
+          files = c("inst/shiny/1"))
+
+zip::zipr(zipfile = "inst/shiny/2/data/demo2.zip", 
+          files = c("inst/shiny/2"))
+