+++ inst/shiny/zip_templates.R
@@ -1,6 +1,0 @@
-zip::zipr(zipfile = "inst/shiny/1/data/template1.zip", 
-          files = c("inst/shiny/1"))
-
-zip::zipr(zipfile = "inst/shiny/2/data/template2.zip", 
-          files = c("inst/shiny/2"))
-